Added standard powershell wildcard support and fixed various issues.
Changed behavior for when a filepath is passed as input to the cmdlet to now write the fileinfo object for those files.

Changed filtering behavior for recursive searches so it searches through all subfolders for matches, instead of only subfolders that matches the filter.

Fixed issue where piping the output to "Select-Object -First X" would cause the cmdlet to continue running through the entire directory tree and spew out warnings.

Fixed issue where the cmdlet would stop for non recursive searches if it encountered an error.
Fixed issue where a relative path would not always be resolved properly.